+++ src/components/app/app.tsx
@@ -3,7 +3,7 @@ import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import MainScreen from '../../pages/main/main';
 import Login from '../../pages/login/login';
 import Favorites from '../../pages/favorites/favorites';
-import Offer from '../../pages/offer/offer';
+import Offer from '../../pages/offer-detailed/offer-detailed';
 import NotFound from '../errors/404';
 import { PrivateRoute } from '../private-route/private-route';
 import { Provider } from 'react-redux';
