+++ src/components/location/location.ts
@@ -1,0 +1,10 @@
+import { City } from '../../types/location';
+
+export const Locations: City[] = [
+  { title: 'Paris', zoom: 7, latitude: 48.868610000000004, longitude: 2.342499},
+  { title: 'Cologne', zoom: 7, latitude: 50.938361, longitude: 6.959974 },
+  { title: 'Brussels', zoom: 7, latitude: 50.846557, longitude: 4.351697 },
+  { title: 'Amsterdam', zoom: 7, latitude: 52.37454, longitude: 4.897976 },
+  { title: 'Hamburg', zoom: 7, latitude: 53.550341, longitude: 10.000654 },
+  { title: 'Dusseldorf', zoom: 7, latitude: 51.225402, longitude: 6.776314 }
+];
