+++ src/components/offer/offer.tsx
@@ -1,9 +1,9 @@
+import React from 'react';
 import { Link } from 'react-router-dom';
 import { OfferProps } from '../../types/offer';
-import { useAppDispatch } from '../../hooks/use-store';
-import { addFavoriteOffer, fetchFavoriteOffers } from '../../store/action';
+import OfferInfo from './offer-info';
 
-function Premium({ isPremium }: OfferProps): false | JSX.Element {
+export function Premium({ isPremium }: OfferProps): false | JSX.Element {
   return (
     isPremium && (
       <div className="place-card__mark">
@@ -13,50 +13,7 @@ function Premium({ isPremium }: OfferProps): false | JSX.Element {
   );
 }
 
-function BookmarkButton({ isFavorite, onClick }: { isFavorite: boolean; onClick: () => void }) {
-  const className = isFavorite ? 'place-card__bookmark-button place-card__bookmark-button--active button' : 'place-card__bookmark-button button';
-  return (
-    <button className={className} type="button" onClick={onClick}>
-      <svg className="place-card__bookmark-icon" width="18" height="19">
-        <use xlinkHref="#icon-bookmark"></use>
-      </svg>
-    </button>
-  );
-}
-
-export function OfferInfo({ offer, offerLink }: { offer: OfferProps; offerLink: string }): JSX.Element {
-  const dispatch = useAppDispatch();
-
-  const onAddOfferToFavorite = () => {
-    dispatch(addFavoriteOffer({ id: offer.id, status: offer.isFavorite ? 0 : 1 }));
-  };
-
-  return (
-    <div className="place-card__info">
-      <div className="place-card__price-wrapper">
-        <div className="place-card__price">
-          <b className="place-card__price-value">&euro;{offer.price}</b>
-          <span className="place-card__price-text">&#47;&nbsp;night</span>
-        </div>
-        <BookmarkButton isFavorite={offer.isFavorite} onClick={onAddOfferToFavorite} />
-      </div>
-      <div className="place-card__rating rating">
-        <div className="place-card__stars rating__stars">
-          <span style={{ width: offer.rating * 20 }}></span>
-          <span className="visually-hidden">Rating</span>
-        </div>
-      </div>
-      <h2 className="place-card__name">
-        <Link to={offerLink}>
-          {offer.title}
-        </Link>
-      </h2>
-      <p className="place-card__type">{offer.type}</p>
-    </div>
-  );
-}
-
-export default function Offer({ offer, setState }: { offer: OfferProps; setState: () => void }): JSX.Element {
+function Offer({ offer, setState }: { offer: OfferProps; setState: () => void }): JSX.Element {
   const offerLink = `/offer/${offer.id}`;
   return (
     <article className="cities__card place-card" onMouseOver={setState}>
@@ -71,48 +28,4 @@ export default function Offer({ offer, setState }: { offer: OfferProps; setState
   );
 }
 
-export function FavoriteOffer({ offer }: { offer: OfferProps }): JSX.Element {
-  const dispatch = useAppDispatch();
-
-  const onRemoveFromFavorite = () => {
-    dispatch(addFavoriteOffer({ id: offer.id, status: 0 }));
-    dispatch(fetchFavoriteOffers());
-  };
-
-  const offerLink = `/offer/${offer.id}`;
-
-  return (
-    <article className="favorites__card place-card">
-      <Premium {...offer} />
-      <div className="favorites__image-wrapper place-card__image-wrapper">
-        <Link to={offerLink}>
-          <img className="place-card__image" src={offer.previewImage} width="150" height="110" alt="Place image" />
-        </Link>
-      </div>
-      <div className="favorites__card-info place-card__info">
-        <div className="place-card__price-wrapper">
-          <div className="place-card__price">
-            <b className="place-card__price-value">&euro;{offer.price}</b>
-            <span
-              className="place-card__price-text"
-            >&#47;&nbsp;night
-            </span>
-          </div>
-          <BookmarkButton isFavorite={offer.isFavorite} onClick={onRemoveFromFavorite} />
-        </div>
-        <div className="place-card__rating rating">
-          <div className="place-card__stars rating__stars">
-            <span style={{ width: offer.rating * 20 }}></span>
-            <span className="visually-hidden">Rating</span>
-          </div>
-        </div>
-        <h2 className="place-card__name">
-          <Link to={offerLink}>
-            {offer.title}
-          </Link>
-        </h2>
-        <p className="place-card__type">{offer.type}</p>
-      </div>
-    </article>
-  );
-}
+export default React.memo(Offer, (prev, next) => prev.offer === next.offer);
