--- src/data/location.ts
@@ -1,10 +1,0 @@
-import { City } from '../types/location';
-
-export const Locations: City[] = [
-  { title: 'Paris', zoom: 7, latitude: 52.3909553943508, longitude: 4.85309666406198},
-  { title: 'Cologne', zoom: 7, latitude: 52.3909553943508, longitude: 4.85309666406198 },
-  { title: 'Brussels', zoom: 7, latitude: 52.3909553943508, longitude: 4.85309666406198 },
-  { title: 'Amsterdam', zoom: 7, latitude: 52.3909553943508, longitude: 4.85309666406198 },
-  { title: 'Hamburg', zoom: 7, latitude: 52.3909553943508, longitude: 4.85309666406198 },
-  { title: 'Dusseldorf', zoom: 7, latitude: 52.3909553943508, longitude: 4.85309666406198 }
-];
