+++ src/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from '../src/components/app/app';
-import { Locations } from './data/location';
+import { Locations } from './components/location/location';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
