+++ src/store/index.ts
@@ -1,7 +1,7 @@
 import { configureStore } from '@reduxjs/toolkit';
 import { reducer } from './reducer';
 import { createAPI } from '../api';
-import { checkAuth } from './action';
+import { checkAuth } from './auth-actions';
 
 export const api = createAPI();
 export const store = configureStore({
