+++ src/types/api.ts
@@ -1,0 +1,10 @@
+export type AddReviewModelApi = {
+    id: string;
+    comment: string;
+    rating: number;
+}
+
+export type AddFavoriteOfferModelApi = {
+  id: string;
+  status: number;
+}
