+++ src/types/store-config.ts
@@ -1,0 +1,9 @@
+import { AxiosInstance } from 'axios';
+import { AppDispatch, State } from '../store/reducer';
+
+
+export type Config = {
+    dispatch: AppDispatch;
+    state: State;
+    extra: AxiosInstance;
+};
